Extract station master data utilities into station.ts
Move fetchStations (extracted from RoadStationMap) and reconcileVisits
into a single station.ts module as both operate on station master data.
Remove reconcile-visits.ts.
